Fix aerobridge credentials token from json file
