+++ tests/fixtures/AerobridgeCredential.json
@@ -6,7 +6,7 @@
         "name": "Chelsea Sims",
         "token_type": 2,
         "association": 1,
-        "token": "2Kkj41m+yB4/RnNiHYCZ4j7jdC2Wblb2pEsTzQBDrkzgdFiSbn4E7TX4mPwTTP//6yXuIDHtr/Jn/3shwnkhRkJLfIn4hDW/iRLgNUiowvKY7FCmEQBKk4UpHSlIxIRetIoOSwHF0F9JDXROIxCF+IrnRbQ3JJKXN6m7UJZDSKvARMdBOg7ARenAgfY9DAG1Lgk4b7Cj9/lAzsQfjx9aM0NKyY2aAoKRQuYLaXJI0fzZYxkG4axnyXWiMN8GfwABD4oqpBkgiOV2+6owf7PpVo6iSB57wIhsCKbhuyerhLNAxytSAM6ef8dUGgjyR028rRFGWRoBWZ9s7FupmmWOwB2E5kth2A9fq1Ea5Ur8Or73OVr8MgBfxMijKeyZsNbX5A+uHquVfTjMnhmJbkJg/fdKAow1X5pGfer7kDHpa7HXQOpY54G8zujWgworYYO5Yf/qHB34n3p5tkVFOaRJ34h9LvRTY1bDH7P8pTIL/7YRgVZyL7b9JAAI5o5xy1xkLXBzT2k2DHvj2XiLjbFziBidcpUwXat1X/8LkdAvGa4CQNGZLnw8MVO5vP2dPYm6gNBiDwoEmj5c9xn0t6dghLp/78bgmiYmAh+ESVnjQChca29m/6blqtUm5gJuRqj+lt1Hn2/goHRFAsaKuTIrDfLluVTJtkqTwHCSepTIgFqakd59Q3KldaBbd+Boqp7T0usPHmRhtcpp8bNlaWSKDejtk9RJD+XFWMU62+/oVBsAUGgZIPkwY9MKmn9v4m9Pt2BRfU8x0Gt6d/Ick6uO7XFKXLXsYSNZv8bL/jfY0aoO7Yiyi60gQqIpNqA5tndSM9/GdI5J0KSP2QUnlxge0YMXPp1fNiBxjhQiVH+dUtNKXCwF8IivnxB2RXvyliWtxjuj7zBkxQysvYR+qmL9FrwP1ZqEWlws2X49DZZpkI/TSC3k8hf5DbhWGWdfDlmQsaKmqpXc5UQrPH5U2d4dWF0UkS9D1xuXy6hAjIENPJ3iRHKq1//P+eRoxQ68t/v9KVv5SgCqsNtJfZVmy6nRxvDmCZIpWAdCxzgGvLsjOmZOu5zz91x+6WCSM42zfecGI8nFeYkfjW+mAV75vK2yiwpFNqUm+YChPv2dI0j+uLrt7TSR2ioyHtdCYh0MsrEV4sYVXD4lrJtqP97qF2fbMpA2AFJabBOBeYHoYRU2MxlIHy14J89r2FVbedUiBaxjFCmq9hPm6b9LVBBF0uLrD0YSv72UwdV2p60ahWO1ZPQiE5FwUeOLNPDH4y/qGERfP/tfhpYAvEG/VZMjfN7vnmL3OxCiNKpQ2dZbNDVwxeicKWadDtLwTcURGeYuHk1kcXmkpNQ4IE6cuhMQG5/3WQ==",
+        "token": "gAAAAABh0fRxOTHRIwKvKwL4bwvpdHIXZNcmHwY2qkoZJirO7AVD4C1-3AdUO1oAM3uCRduc0mmGr2skBu7Iil6fZn3JQhJmWIZaP2U6dYPCQUGzyfJBPkiKCjV8D60LHhQVa49qbds7xXUSe9jMSOt9i5Jw6xvVscA4qeVIDZEjO68R8AyXl83ZcFKRXW8mDtw7Cowp9XrwNCgprP1JG_OHoOQ-uZVW5KtTTo9wml-_gVN2EaY5exHkUhKsmjNeAkHzzHKUXOW6_lcBoZrW1SkYY5cJcid3pP4O2_2MUmUTyHujPttwh5QasFiQdvbhscLykBaRdZ5zkcq8EyfVrYJH8nnyHxWAorPK3s6XOI8LpccgTlvirYStzzHdLjzaLeHvv8KS1do9nMeKVRmYrhg2u_8OabRupcADq6QU64LdHGnb9CyOCXXK22jCOhRffoaIW7cs861z1SpKgH9H7nA9bfkhG8BOzYiqfOMNxBYQP7HCeO1pFY7s-ffojH1fP0ItDeauAoRCi3lj36nyyZILGoN12ZCY6RTZaG-W_h7KgVbRPBwrEmlUracuiiYBGMccHsMq0nLVMSuZ29Oft6ktAvscQKSdwpS8EwVHZWL3iSWhAfWTJMGUgXYRGVGMSiEcVIz61rUzG9MPPRkhDRJzZDLnWGsFOwOrJDs5E8F28FSOs9oLrHufQw2Z9cXUAz7WzzsJ67BUeujruCdJHIaeLlMquwjAw8IWidET2sbgpq5RFH5GS0Ht_tR6zGWFSf9j3S6AGeSngz8jAjg9MoFwDGztDZU5CJmrjuVAju4XRhLdUSLLJr_Ip-qsseAG2fUPVnU2U_6kHXBgzEunp-K0y2SfCJPxjZYrh9q-WxDePxAke9KXdHtYE4VufXvw9XWvf6mh8OsLr9QVdmdX1H5aiWUz6Wx7wGh2PY7uH29CfArtmHGUHT0he_Vn1kNlWH3SVOkXkg2iAwi2gDOokz2y1LmdzlpDs8cUhEs_aPwsrjilV5uH3XpFOD145VfpvSRHuKwu8sYXPtfGrGhEdEjX_-z_194I8xyoEzGzykxlo8uK8N6kC0KyL2Fdx0WCTnvVe2EGJWpZCDsYterBuuRRBGuqOsp65tymb0TRBc2LqgllIUcLVxQR_tVl6P7qaCmcetv-I-mOeke2SUox3zXZQ-T5kN0Tm0dFQqCY6nn7tmLbaBUJ8g1zKOaFyGWDNbEre2bdNTT0OBN1Vir7mSuf3VJpfTZz17mGCv1-LQLU7kTmCS6etIU=",
         "aircraft": null,
         "manufacturer": "31fd7ede-c617-4f8a-8bb8-a69e304a21e9",
         "operator": null,
@@ -22,7 +22,7 @@
         "name": "Marie Young",
         "token_type": 1,
         "association": 0,
-        "token": "2Kkj41m+yB4/RnNiHYCZ4j7jdC2Wblb2pEsTzQBDrkzgdFiSbn4E7TX4mPwTTP//6yXuIDHtr/Jn/3shwnkhRkJLfIn4hDW/iRLgNUiowvKY7FCmEQBKk4UpHSlIxIRetIoOSwHF0F9JDXROIxCF+IrnRbQ3JJKXN6m7UJZDSKvARMdBOg7ARenAgfY9DAG1Lgk4b7Cj9/lAzsQfjx9aM0NKyY2aAoKRQuYLaXJI0fzZYxkG4axnyXWiMN8GfwABD4oqpBkgiOV2+6owf7PpVo6iSB57wIhsCKbhuyerhLNAxytSAM6ef8dUGgjyR028rRFGWRoBWZ9s7FupmmWOwB2E5kth2A9fq1Ea5Ur8Or73OVr8MgBfxMijKeyZsNbX5A+uHquVfTjMnhmJbkJg/fdKAow1X5pGfer7kDHpa7HXQOpY54G8zujWgworYYO5Yf/qHB34n3p5tkVFOaRJ34h9LvRTY1bDH7P8pTIL/7YRgVZyL7b9JAAI5o5xy1xkLXBzT2k2DHvj2XiLjbFziBidcpUwXat1X/8LkdAvGa4CQNGZLnw8MVO5vP2dPYm6gNBiDwoEmj5c9xn0t6dghLp/78bgmiYmAh+ESVnjQChca29m/6blqtUm5gJuRqj+lt1Hn2/goHRFAsaKuTIrDfLluVTJtkqTwHCSepTIgFqakd59Q3KldaBbd+Boqp7T0usPHmRhtcpp8bNlaWSKDejtk9RJD+XFWMU62+/oVBsAUGgZIPkwY9MKmn9v4m9Pt2BRfU8x0Gt6d/Ick6uO7XFKXLXsYSNZv8bL/jfY0aoO7Yiyi60gQqIpNqA5tndSM9/GdI5J0KSP2QUnlxge0YMXPp1fNiBxjhQiVH+dUtNKXCwF8IivnxB2RXvyliWtxjuj7zBkxQysvYR+qmL9FrwP1ZqEWlws2X49DZZpkI/TSC3k8hf5DbhWGWdfDlmQsaKmqpXc5UQrPH5U2d4dWF0UkS9D1xuXy6hAjIENPJ3iRHKq1//P+eRoxQ68t/v9KVv5SgCqsNtJfZVmy6nRxvDmCZIpWAdCxzgGvLsjOmZOu5zz91x+6WCSM42zfecGI8nFeYkfjW+mAV75vK2yiwpFNqUm+YChPv2dI0j+uLrt7TSR2ioyHtdCYh0MsrEV4sYVXD4lrJtqP97qF2fbMpA2AFJabBOBeYHoYRU2MxlIHy14J89r2FVbedUiBaxjFCmq9hPm6b9LVBBF0uLrD0YSv72UwdV2p60ahWO1ZPQiE5FwUeOLNPDH4y/qGERfP/tfhpYAvEG/VZMjfN7vnmL3OxCiNKpQ2dZbNDVwxeicKWadDtLwTcURGeYuHk1kcXmkpNQ4IE6cuhMQG5/3WQ==",
+        "token": "gAAAAABh0ffA5nud35Y8MmV2sxzCCiFpgT5InEStFZHhL5u2t-nMCUO3MSZVSM76oRQvIh0ZWj_qsTdL-OzwRPUcT-2VAaPOsdPMBeg0ZLtWOuertWhcG-IK9z1UahWZ7BAFDmjqPQZdlbWqUni32pCALT86pedWcxn83ZetB-_ApHPQx-epc3GV4D8NStDUI9McFTXQMrkZ5Ue_zwhVCsiNhYljHTZpU7DyjrP5Jk9iSfYOvME0lFGr9QxiczE22VYiXW0FVBGtQ_tRJcZLItdCWeCEG88KenB5dBTjihpYTf8c_z8ddt4hSZMcxI4NQnXDqqaOCUHBTqWxsipGWwAkyL4rgC-H7WCAsCOB76_R2AFz1VBdM4wfACG_JQ-0NjeKOnlE0G4OavF9FAzD2WkizNUBGbcJfcQ-sxjonHPe4krt5XIdLah2CCbqlT_2bQcLB0p5GBRSKHuuH8xlupBbMTyoithEaXvVkCsauc8fPrKHe1HTf1hrLMVErGi3YNGIBAbvLZXFbRO-uF1rgkVOgwsSSS6MY6bDF5vNKoBh1MHUBphUXLSAXUsmSKsPBU0I0Hki3IunjTmgn0_92FmlywRDgENv--oN8k7rL_sIKpan0AhuqNOqoXZ_FSQCmNMuM2vxKY8L8sr0uwehcpADOexf8-bWndVQP1mbWKJqyVoKvdFtn-c7AgQ8fLv7mnthcxDLn1q7JmiyOPL_kK8h8cnCfIJ1uD7gXwZTQiGpRZGFl6NBrQ6OxSyJd7QVx-birCsBrzbIAPZWJmxAIfuwj_28J_E29l7R2mks6c9wVkyxJMrNQ_Mer6f3SQLNAwo75k-EnrVyOwpVpVQEARpZhUcrKsv8LnwFLlb-Z_dVeCW0jWWc5gAYcoKAAUY0l3JbYV1LDFvWzBnEZMUj1Z_z6sY5yADNUi4yV8dAnGh3_vw8rWbcCoTD7raB7O7nZpZpjf8LAawPph49Gw9xtkp4FNwnuLETyUJKcWcHBVZ-jdhn7TclgiRd3DHvBqscFJKpb8WcbQi_W0MhcOCK7J5ayMFbxvokTbem-1x2HmD5KnBeSwdT64-w3vM2w04MXnplCXv_k9R2rPXruRA9EDeiwJHsJBHB6G2dSyStdoi-61m1tMbGJXiojQte3AtA3KeYvSX6DyPYcdYNGSOdHppRI75gMf4xvRA7YRiWCiM2mwdeZM9ognQn91jz0_1DX7k-UZ2XD0cjdwl0d2n8HoJxX2hdLt1LPut9dgATpd_WgqNlURaEXZ4=",
         "aircraft": null,
         "manufacturer": null,
         "operator": "66ff7549-5737-495d-965b-2d93c0bc5b5e",
