+++ tests/test_restApis/test_flight_logs.py
@@ -10,6 +10,7 @@ class TestFlightLogs(TestApiEndpoints):
                 'SignedFlightLog']
 
     def setUp(self):
+        self.fix_fixtures_data()
         self.setUpClientCredentials([self.READ_SCOPE, self.WRITE_SCOPE])
 
     def test_flight_log_list_get_returns_200(self):
@@ -90,7 +91,6 @@ class TestFlightLogs(TestApiEndpoints):
         self.assertEqual(res.json(), {'message': 'No flight Log found to sign'})
 
     def test_log_sign_put_returns_409(self):
-        self.fixAerobridgeCredentialData()
         url = reverse('log-sign', kwargs={'pk': self.get_pk_for_model('FlightLog')})
 
         res = self.client.put(url)
@@ -99,7 +99,6 @@ class TestFlightLogs(TestApiEndpoints):
 
     # TODO: Include this test when bug in sign_log fixed
     def exclude_test_log_sign_put_returns_200(self):
-        self.fixAerobridgeCredentialData()
         url = reverse('log-sign', kwargs={'pk': self.get_pk_for_model('FlightLog', 1)})
 
         required_keys = {'id', 'signed_log', 'created_at', 'updated_at', 'raw_flight_log'}
